+++ src/utils.cpp
@@ -36,7 +36,7 @@ fnc Utils::ReadFile(const char* path) -> char* {
   return buffer;
 }
 
-fnc Utils::HashString(const char* str, u32 length) -> u32 {
+fnc Utils::HashString(const char* str, u64 length) -> u32 {
   u32 hash = 2166136261u;
 
   for (u32 i = 0; i < length; i++) {
